\name{Toc-methods}

\docType{methods}

\alias{Roc-method}
\alias{Toc-method}

\alias{setTable}
\alias{setTable, Roc-method}
\alias{setTable, Toc-method}
\alias{setAUC}
\alias{setAUC, Roc-method}
\alias{setAUC, Toc-method}
\alias{setMaxAUC}
\alias{setMaxAUC, Roc-method}
\alias{setMaxAUC, Toc-method}
\alias{setMinAUC}
\alias{setMinAUC, Roc-method}
\alias{setMinAUC, Toc-method}
\alias{setPrevalence}
\alias{setPrevalence, Toc-method}
\alias{setPopulation}
\alias{setPopulation, Toc-method}
\alias{setUnits}
\alias{setUnits, Toc-method}

\alias{setTable, Roc}
\alias{setTable, Toc}
\alias{setAUC, Roc}
\alias{setAUC, Toc}
\alias{setMaxAUC, Roc}
\alias{setMaxAUC, Toc}
\alias{setMinAUC, Roc}
\alias{setMinAUC, Toc}
\alias{setPrevalence, Toc}
\alias{setPopulation, Toc}
\alias{setUnits, Toc}

\title{Methods for the Roc and Toc classes}

\description{
Methods for the Roc and Toc classes
}

\details{
Methods for the Roc and Toc classes
}


\value{
an object of class Toc or Roc
}
\seealso{
\code{\link{TOC}}, \code{\link{ROC}}
}

\keyword{methods}
\keyword{spatial}

