\name{TOC-package}
\alias{TOC-package}
\docType{package}
\title{
Total Operating Characteristic (TOC) Curve and ROC Curve
}
\description{
Construction of the Total Operating Characteristic (TOC) Curve and the Receiver (aka Relative) Operating Characteristic (ROC) Curve for spatial and non-spatial data.}
\details{
\tabular{ll}{
Package: \tab TOC\cr
Type: \tab Package\cr
Version: \tab 0.0-2\cr
Date: \tab 2015-03-28\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Robert G. Pontius <rpontius@clarku.edu>, Ali Santacruz, Amin Tayyebi, Benoit Parmentier, Kangping Si

Maintainer: Ali Santacruz <amsantac@unal.edu.co>
}
\keyword{ package }
\keyword{ spatial }
\seealso{
\code{\link{TOC}}, \code{\link{plot}} 
}
