\name{TOC-package}
\alias{TOC-package}
\alias{TOC-package}
\docType{package}
\title{
Construct the Total Operating Characteristic (TOC) Curve
}
\description{
Construct the Total Operating Characteristic (TOC) Curve}
\details{
\tabular{ll}{
Package: \tab TOC\cr
Type: \tab Package\cr
Version: \tab 0.0-1\cr
Date: \tab 2014-10-29\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Robert G. Pontius <rpontius@clarku.edu>, Alí Santacruz, Amin Tayyebi, Benoit Parmentier, Kangping Si

Maintainer: Alí Santacruz <asantacruzdelgado@clarku.edu>
}
\keyword{ package }
\keyword{ spatial }
\seealso{
\code{\link{TOC}}, \code{\link{plot.TOC}} 
}
